Rearrange specs and deployment config for travis jobs
We want to run the full test suite and then deploy from that same job for staging, so that we get JS coverage stats uploaded live on staging.

While we're at it, this will remove almost all the Ruby specs from Job #1 so that production deployments happen faster. This assumes that you should only be pushing to production when the build has already passed on the main branch.